scripts/python/Get-chrome-pass/requirements.txt
pytest-shutil
pycrypto
pypiwin32
pysqlite3
pybase64
jsonlib
os-sys